refactor: build context
