+++ packages/contract/hardhat.config.ts
@@ -1,9 +1,10 @@
 import * as dotenv from 'dotenv'
 
 import { HardhatUserConfig, task } from 'hardhat/config'
+import '@typechain/hardhat'
 import '@nomiclabs/hardhat-etherscan'
 import '@nomiclabs/hardhat-waffle'
-import '@typechain/hardhat'
+import '@nomiclabs/hardhat-ethers'
 import 'hardhat-gas-reporter'
 import 'solidity-coverage'
 
@@ -36,6 +37,10 @@ const config: HardhatUserConfig = {
   },
   etherscan: {
     apiKey: process.env.ETHERSCAN_API_KEY
+  },
+  typechain: {
+    outDir: 'src/generated',
+    target: 'ethers-v5'
   }
 }
