+++ packages/contract/package.json
@@ -10,7 +10,7 @@
     "build": "run-p -l build:*",
     "build:cjs": "tsc --project . --module commonjs --outDir ./cjs",
     "build:esm": "tsc --project . --module es2015 --outDir ./esm",
-    "clean": "rimraf esm cjs",
+    "clean": "rimraf esm cjs artifacts",
     "start:node": "hardhat node",
     "console": "hardhat console",
     "compile": "hardhat compile",
